Conflicts:
	Euler/Problems/Problem52.cs

+++ Euler/Problems/Problem43.cs
@@ -11,7 +11,7 @@ namespace Euler.Problems {
 
 		public override string Solve() {
 			var generator = new PermutationGenerator<string>(new List<string> { "0", "1", "2", "3", "4", "5", "6", "7", "8", "9" });
-			BigInteger sum = 0;
+			long sum = 0;
 			while (generator.HasMore) {
 			  var perm = generator.GetNext();
 
@@ -21,7 +21,7 @@ namespace Euler.Problems {
 			return sum.ToString();
 		}
 
-		BigInteger HasSpecialProperty(List<string> numberString) {
+		long HasSpecialProperty(List<string> numberString) {
 			var index = 1;
 			bool wasDivisible = true;
 			foreach (var divisor in divisors) {
@@ -32,7 +32,7 @@ namespace Euler.Problems {
 				var first = numberString[index];
 				var second = numberString[index + 1];
 				var third = numberString[index + 2];
-				var number = BigInteger.Parse(first + second + third);
+				var number = long.Parse(first + second + third);
 
 				if ((number % divisor) != 0) {
 					wasDivisible = false;
@@ -41,7 +41,7 @@ namespace Euler.Problems {
 				index++;
 			}
 
-			return wasDivisible ? BigInteger.Parse(string.Join("", numberString)) : 0;
+			return wasDivisible ? long.Parse(string.Join("", numberString)) : 0;
 		}
 
 		public override string Solution {
